requirements.txt
flask
scikit-learn
pandas
datasets
fasttext-wheel
numpy<2
